[Concurrency][WIP] CancellationError.Reason + cancel(reason:) + cancellationReason + hasActiveDeadline + activeDeadline(for:)
Introduce the reason-plumbing pieces from SE-0526:

- CancellationError.Reason (@nonexhaustive-ish enum: .unspecified, .deadlineExpired)
- CancellationError.init(reason:) + .reason property (backed by a UInt8
  raw-storage field on the non-@frozen struct)
- Task.cancel(reason:), UnsafeCurrentTask.cancel(reason:)
- Task.cancellationReason (static, on Task where Success==Never, Failure==Never)
- UnsafeCurrentTask.cancellationReason
- TaskGroup.cancelAll(reason:), ThrowingTaskGroup, DiscardingTaskGroup,
  ThrowingDiscardingTaskGroup - all four variants
- Task.hasActiveDeadline, Task.activeDeadline(for:)

Runtime plumbing:
- ActiveTaskStatus: new CancelReasonDeadlineExpired = 0x2000000 flag qualifying
  IsCancelled. First-cancel-wins.
- withCancelled(size_t reason) and getCancellationReason() accessors.
- swift_task_cancelWithReason(task, reason) entry point via compat-override;
  swift_task_cancel delegates to it with reason=0.
- Reason threaded through performCancellationAction and \_swift_taskGroup_cancel*
  so child-task recursion inherits the parent's reason.
- swift_task_getCancellationReason(task) fast-path reader.
- swift_taskGroup_cancelAllWithReason.
- _swift_task_hasActiveDeadline() fast-path Swift bridge.

This commit is intentionally build-broken on the pending SIL ownership
issue in _findNearestDeadline (leaked owned Builtin.NativeObject) that
the next commit will resolve as part of the record redesign. The
Reason plumbing is complete and orthogonal to the record layout.

+++ stdlib/public/Concurrency/Clock.swift
@@ -140,7 +140,7 @@ enum _ClockID: Int32 {
 /// deadline runtime distinguishes system clocks from custom clocks via a
 /// tagged discriminator, not the numeric value.
 @available(StdlibDeploymentTarget 6.5, *)
-public enum SystemClockID: Int64, Sendable, Hashable {
+public enum SystemClockID: UInt8, Sendable, Hashable {
   case continuous = 1
   case suspending = 2
   case walltime = 3
